refactor(common): 단일 도메인 전용 컴포넌트를 해당 도메인으로 이동 및 미사용 코드 삭제
- If.tsx → challenge/ui/, RequiredStar/DrawerCloseBtn → report/ui/
- ReportSubmitModal → report/modal/, MarketingConsentSection → library/ui/
- MobileCarousel → mypage/ui/
- 미사용 placeholder/, scroll-to-top/ 폴더 삭제
- 빈 drawer/, carousel/, form/ 폴더 정리
