--- apps/web/src/common/placeholder/CardListPlaceholder.module.scss
@@ -1,8 +1,0 @@
-.placeholder {
-  width: 100%;
-  height: 18.75rem;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  color: #4a495c;
-}
