+++ apps/web/src/domain/mypage/ui/MobileCarousel.tsx
