+++ apps/web/src/domain/report/ui/DrawerCloseBtn.tsx
