+++ apps/web/src/domain/report/ui/RequiredStar.tsx
